tests/wpt/web-platform-tests/css/css-backgrounds/parsing/background-clip-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Backgrounds and Borders Module Level 3: parsing background-clip with valid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-backgrounds/#background-clip">
<meta name="assert" content="background-clip supports the full grammar '<box>#'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("background-clip", "border-box");
test_valid_value("background-clip", "padding-box");
test_valid_value("background-clip", "content-box");

test_valid_value("background-clip", "border-box, padding-box, content-box");
</script>
</body>
</html>
